replica shutdown sequence for direct ingress (#1902)
**New Env Var: RAY_SERVE_DISABLE_SHUTTING_DOWN_REPLICAS_FORCEFULLY**
Introduced this instead of hardcoding logic tied to
ANYSCALE_RAY_SERVE_ENABLE_DIRECT_INGRESS. This keeps the behavior
decoupled from Direct Ingress and allows OSS users to opt into it
independently.

**Skip Forceful Shutdown Instead of Extending Timeout**
Decided not to add a configurable timeout for forceful shutdowns, as the
proxy currently doesn’t support timeout functionality either. This
maintains consistency across components.

**Avoid Reusing graceful_shutdown_timeout_s**
This config is used for selecting replicas pending migration and for
timing graceful shutdowns. Reusing it here could introduce unintended
behavior, so a separate toggle was preferred.

**Net Effect**
To enable Direct Ingress with this change, users must set both:

1. RAY_SERVE_DISABLE_SHUTTING_DOWN_REPLICAS_FORCEFULLY=1

reference doc
https://docs.google.com/document/d/1kOd2-ADwHKl5MGBnYV6kveuDoMMY7bdaygGmogfp3Qw/edit?tab=t.0

---------

+++ python/ray/serve/_private/deployment_state.py
@@ -73,6 +73,13 @@ from ray.serve.schema import (
 )
 from ray.util.placement_group import PlacementGroup
 
+# isort: off
+from ray.anyscale.serve._private.constants import (
+    ANYSCALE_RAY_SERVE_ENABLE_DIRECT_INGRESS,
+)
+
+# isort: on
+
 logger = logging.getLogger(SERVE_LOGGER_NAME)
 
 
@@ -187,6 +194,20 @@ SLOW_STARTUP_WARNING_PERIOD_S = int(
 
 ALL_REPLICA_STATES = list(ReplicaState)
 _SCALING_LOG_ENABLED = os.environ.get("SERVE_ENABLE_SCALING_LOG", "0") != "0"
+# Feature flag to disable forcibly shutting down replicas.
+RAY_SERVE_DISABLE_SHUTTING_DOWN_INGRESS_REPLICAS_FORCEFULLY = (
+    os.environ.get("RAY_SERVE_DISABLE_SHUTTING_DOWN_INGRESS_REPLICAS_FORCEFULLY", "0")
+    == "1"
+)
+if (
+    ANYSCALE_RAY_SERVE_ENABLE_DIRECT_INGRESS
+    and not RAY_SERVE_DISABLE_SHUTTING_DOWN_INGRESS_REPLICAS_FORCEFULLY
+):
+    logger.info(
+        "Setting RAY_SERVE_DISABLE_SHUTTING_DOWN_INGRESS_REPLICAS_FORCEFULLY to True "
+        "because ANYSCALE_RAY_SERVE_ENABLE_DIRECT_INGRESS is set to True."
+    )
+    RAY_SERVE_DISABLE_SHUTTING_DOWN_INGRESS_REPLICAS_FORCEFULLY = True
 
 
 def print_verbose_scaling_log():
@@ -266,6 +287,7 @@ class ActorReplicaWrapper:
         self._routing_stats: Dict[str, Any] = {}
         self._record_routing_stats_ref: Optional[ObjectRef] = None
         self._last_record_routing_stats_time: float = 0.0
+        self._ingress: bool = False
 
     @property
     def replica_id(self) -> str:
@@ -414,6 +436,7 @@ class ActorReplicaWrapper:
         until the deployment scheduler schedules the underlying actor.
         """
         self._actor_resources = deployment_info.replica_config.resource_dict
+        self._ingress = deployment_info.ingress
         # it is currently not possible to create a placement group
         # with no resources (https://github.com/ray-project/ray/issues/20401)
         self._deployment_is_cross_language = (
@@ -972,6 +995,17 @@ class ActorReplicaWrapper:
 
     def force_stop(self):
         """Force the actor to exit without shutting down gracefully."""
+        if (
+            self._ingress
+            and RAY_SERVE_DISABLE_SHUTTING_DOWN_INGRESS_REPLICAS_FORCEFULLY
+        ):
+            logger.info(
+                f"{self.replica_id} did not shut down because it had not finished draining requests. "
+                "Going to wait until the draining is complete. You can force-stop the replica by "
+                "setting RAY_SERVE_DISABLE_SHUTTING_DOWN_INGRESS_REPLICAS_FORCEFULLY to 0."
+            )
+            return
+
         try:
             ray.kill(ray.get_actor(self._actor_name, namespace=SERVE_NAMESPACE))
         except ValueError:
